Complete creating order

+++ main/java/ru/carwash/dto/Service.kt
@@ -1,4 +1,4 @@
 package ru.carwash.dto
 
 /* Услуга автомойки */
-class Service(var name: String, var price: Int)
+class Service(var name: String, var price: Double)
